Many changes! Simple Lambert light system, other testings, new array3 non-optimized memory type... test octree non completed

+++ src/voxels/particlesmodel.cpp
@@ -9,7 +9,6 @@ void ParticlesModel::addVoxel(Voxel* voxel) {
     _voxels.push_back(voxel);
 }
 
-
 size_t ParticlesModel::getVoxelsCount() {
     return _voxels.size();
 }
@@ -21,7 +20,7 @@ void ParticlesModel::setVoxel(int num, Voxel* voxel) {
 Voxel* ParticlesModel::getVoxel(int num) {
     return _voxels[num];
 }
-Voxel* ParticlesModel::getVoxel(glm::vec3 position) {
+Voxel* ParticlesModel::getVoxel(glm::ivec3 position) {
     for (size_t i = 0; i < _voxels.size(); i++) {
         if (_voxels[i]->getPosition() == position) {
             return _voxels[i];
@@ -39,9 +38,9 @@ glm::vec3 ParticlesModel::getSize() {
 }
 
 
-void ParticlesModel::setVoxelPosition(int num, glm::vec3 position) {
+void ParticlesModel::setVoxelPosition(int num, glm::ivec3 position) {
     _voxels[num]->setPosition(position);
 }
-glm::vec3 ParticlesModel::getVoxelPosition(int num) {
+glm::ivec3 ParticlesModel::getVoxelPosition(int num) {
     return _voxels[num]->getPosition();
 }
